refactor(billing): replace BullMQ worker + Redis with Coolify scheduled scripts
Billing state (nextChargeAt, failedAttempts/dunning) already lives in the DB, so
recurring charges only need a cron trigger — not a long-running worker + Redis.
Coolify's native Scheduled Tasks run two one-shot CLI scripts in the app
container instead.

- Add src/scripts/charge-due-subscriptions.ts + refresh-square-tokens.ts
  (run once, exit); package.json scripts `charge-due` + `refresh-tokens`.
- Remove src/worker/*, src/lib/jobs/*, bullmq + ioredis + tsup; add tsx as a
  production dependency (scripts run via tsx in the image).
- docker-compose: drop the redis service. .env.example: drop REDIS_URL.
- email/{resend,payment-receipt}: use the bare 'server-only' directive so the
  scripts run under tsx (matches the rest of the lib).
- docs: Coolify Scheduled Tasks instead of a worker + Redis.
