+++ docker-compose.yml
@@ -18,22 +18,6 @@ services:
       timeout: 5s
       retries: 5
 
-  # Redis backs the BullMQ worker (Square recurring charges). Run the worker
-  # locally with: REDIS_URL=redis://localhost:6379 npm run worker
-  redis:
-    image: redis:7-alpine
-    restart: always
-    command: ["redis-server", "--appendonly", "yes"]
-    ports:
-      - '6379:6379'
-    volumes:
-      - redis:/data
-    healthcheck:
-      test: ["CMD", "redis-cli", "ping"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
   # Optional: pgAdmin for database management
   pgadmin:
     image: dpage/pgadmin4:latest
@@ -51,8 +35,6 @@ services:
 
 volumes:
   db:
-    driver: local
-  redis:
     driver: local
   pgadmin:
     driver: local
