+++ package.json
@@ -26,8 +26,8 @@
     "db:studio": "prisma studio",
     "db:migrate:prod": "npx prisma migrate deploy",
     "db:reset": "prisma migrate reset",
-    "worker": "tsx src/worker/worker-main.ts",
-    "worker:dev": "tsx watch src/worker/worker-main.ts",
+    "charge-due": "tsx src/scripts/charge-due-subscriptions.ts",
+    "refresh-tokens": "tsx src/scripts/refresh-square-tokens.ts",
     "prod_start": "npm run db:migrate:prod && npm run start",
     "test": "vitest run",
     "test:watch": "vitest"
@@ -79,7 +79,6 @@
     "@uiw/react-codemirror": "^4.25.10",
     "ai": "^6.0.194",
     "archiver": "^7.0.1",
-    "bullmq": "^5.78.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -87,7 +86,6 @@
     "dotenv": "^16.5.0",
     "embla-carousel-react": "^8.6.0",
     "exceljs": "^4.4.0",
-    "ioredis": "^5.11.1",
     "jspdf": "^4.2.1",
     "lucide-react": "^0.475.0",
     "nanoid": "^5.1.7",
@@ -108,6 +106,7 @@
     "square": "^44.1.0",
     "tailwind-merge": "^3.0.1",
     "tailwindcss-animate": "^1.0.7",
+    "tsx": "^4.22.4",
     "zod": "^3.25.76"
   },
   "devDependencies": {
@@ -121,7 +120,6 @@
     "postcss": "^8.5.12",
     "prettier": "^3.8.1",
     "tailwindcss": "^3.4.1",
-    "tsup": "^8.5.1",
     "typescript": "5.9.3",
     "vitest": "^4.0.18"
   },
